- optimise path resolving process
