+++ srv/configure.js
@@ -5,21 +5,7 @@ const config = require('./config.json')
 const express = require('express')
 const builder = require('./builder')
 const args = minimist(process.argv)
-const definePath = function (directPath, defaultPath, entity) {
-  if (directPath) {
-    return path.resolve(directPath)
-  }
-  return path.join(path.resolve(defaultPath), entity)
-}
-const defineRootPath = function (embedded) {
-  if (embedded) {
-    if (args.feRootPath) {
-      return path.resolve(args.feRootPath)
-    }
-    return path.resolve('../../')
-  }
-  return path.resolve(__dirname, '../cms.frontend')
-}
+const { defineRootPath, definePath } = require('./path')
 const rootPath = defineRootPath(args.embedded)
 const componentsPath = definePath(args.componentsPath, rootPath, config.components)
 const pagesPath = definePath(args.pagesPath, rootPath, config.pages)
