+++ vue.config.js
@@ -1,24 +1,9 @@
 const configureAPI = require('./srv/configure')
 const minimist = require('minimist')
-const path = require('path')
 const config = require('./srv/config.json')
+const { defineRootPath, definePath } = require('./srv/path')
 
 const args = minimist(process.argv)
-const definePath = function (directPath, defaultPath, entity) {
-  if (directPath) {
-    return path.resolve(directPath)
-  }
-  return path.join(path.resolve(defaultPath), entity)
-}
-const defineRootPath = function (embedded) {
-  if (embedded) {
-    if (args.feRootPath) {
-      return path.resolve(args.feRootPath)
-    }
-    return path.resolve('../../')
-  }
-  return path.resolve(__dirname, 'cms.frontend')
-}
 const rootPath = defineRootPath(args.embedded)
 const componentsPath = definePath(args.componentsPath, rootPath, config.components)
 const pagesPath = definePath(args.pagesPath, rootPath, config.pages)
